Allow deleting events.
Note that eventIDs are not stable. When we delete an event, greater eventIDs
decrement. This is an artefact of using arrays in localStorage as our database.
I'd prefer a relational database, but this is just an exercise so I'm playing
it out.

This fixes a bug in updateEventDates(), where eventIDs were not being removed
from dates due to a type issue. JS is quite forgiving of numeric strings,
sometimes. You can use a numeric string in square bracket notation to reference
an array index. You can compare a numeric string to a number using
inequalities. However, indexOf will not find an integer when given that same
integer as a string. TypeScript is looking pretty good right about now.