Include exact short hex codes for color names in color substrings.
Previously, if you filter colors with a short hex code (e.g. "#639"), you would
not see the color name that matches ("rebeccapurple" with code "#663399"). This
change fixes that, so you can see color names when you filter by a 3 digit code
that corresponds exactly to the 6 digit hex code for a color name.